Provide environment variables for shell calls
There is a string slice for the command available. The entries
are Environment variables. Currently it is (only) possible to
use the environment from the caller. Now it is possible to set new
environment variables or to overwrite existing environment variables.

Can be used e.g. for re-defining HOME or for defining additional
tool specific environment variables.